+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- hw-ledger, keyring, util, util-crypto 10.1.11
 - api, api-contract, types 9.5.1
 - react-{identicon, qr}, ui-{keyring, settings}, vue-identicon 2.9.11
 - phishing 0.18.10